refactor: CustomBoardFactory test 디렉토리로 이동
